[TEST] move testBulkRequestModifier to existing BulkRequestModifierTests class
